+++ src/types/topic.ts
@@ -2,6 +2,7 @@ import { BaseDataModel } from '@/types/meta';
 
 export interface ChatTopic extends Omit<BaseDataModel, 'meta'> {
   favorite?: boolean;
+  sessionId?: string;
   title: string;
 }